Move some album resolvers to actions, refactor album tests
- Album resolvers has been refactored to make it easier to test, this is
done by converting some of the resolvers to actions.

- The album tests has been rewritten (and simplified) to accommodate the changes.
