+++ api/graphql/models/actions/album_actions.go
@@ -1,0 +1,124 @@
+package actions
+
+import (
+	"github.com/photoview/photoview/api/graphql/models"
+	"github.com/pkg/errors"
+	"gorm.io/gorm"
+)
+
+func MyAlbums(db *gorm.DB, user *models.User, order *models.Ordering, paginate *models.Pagination, onlyRoot *bool, showEmpty *bool, onlyWithFavorites *bool) ([]*models.Album, error) {
+	if err := user.FillAlbums(db); err != nil {
+		return nil, err
+	}
+
+	if len(user.Albums) == 0 {
+		return nil, nil
+	}
+
+	userAlbumIDs := make([]int, len(user.Albums))
+	for i, album := range user.Albums {
+		userAlbumIDs[i] = album.ID
+	}
+
+	query := db.Model(models.Album{}).Where("id IN (?)", userAlbumIDs)
+
+	if onlyRoot != nil && *onlyRoot {
+		query = query.Where("parent_album_id IS NULL")
+	}
+
+	if showEmpty == nil || !*showEmpty {
+		subQuery := db.Model(&models.Media{}).Where("album_id = albums.id")
+
+		if onlyWithFavorites != nil && *onlyWithFavorites {
+			favoritesSubquery := db.
+				Model(&models.UserMediaData{UserID: user.ID}).
+				Where("user_media_data.media_id = media.id").
+				Where("user_media_data.favorite = true")
+
+			subQuery = subQuery.Where("EXISTS (?)", favoritesSubquery)
+		}
+
+		query = query.Where("EXISTS (?)", subQuery)
+	}
+
+	query = models.FormatSQL(query, order, paginate)
+
+	var albums []*models.Album
+	if err := query.Find(&albums).Error; err != nil {
+		return nil, err
+	}
+
+	return albums, nil
+}
+
+func Album(db *gorm.DB, user *models.User, id int) (*models.Album, error) {
+	var album models.Album
+	if err := db.First(&album, id).Error; err != nil {
+		if errors.Is(err, gorm.ErrRecordNotFound) {
+			return nil, errors.New("album not found")
+		}
+		return nil, err
+	}
+
+	ownsAlbum, err := user.OwnsAlbum(db, &album)
+	if err != nil {
+		return nil, err
+	}
+
+	if !ownsAlbum {
+		return nil, errors.New("forbidden")
+	}
+
+	return &album, nil
+}
+
+func SetAlbumCover(db *gorm.DB, user *models.User, mediaID int) (*models.Album, error) {
+	var media models.Media
+
+	if err := db.Find(&media, mediaID).Error; err != nil {
+		return nil, err
+	}
+
+	var album models.Album
+
+	if err := db.Find(&album, &media.AlbumID).Error; err != nil {
+		return nil, err
+	}
+
+	ownsAlbum, err := user.OwnsAlbum(db, &album)
+	if err != nil {
+		return nil, err
+	}
+
+	if !ownsAlbum {
+		return nil, errors.New("forbidden")
+	}
+
+	if err := db.Model(&album).Update("cover_id", mediaID).Error; err != nil {
+		return nil, err
+	}
+
+	return &album, nil
+}
+
+func ResetAlbumCover(db *gorm.DB, user *models.User, albumID int) (*models.Album, error) {
+	var album models.Album
+	if err := db.Find(&album, albumID).Error; err != nil {
+		return nil, err
+	}
+
+	ownsAlbum, err := user.OwnsAlbum(db, &album)
+	if err != nil {
+		return nil, err
+	}
+
+	if !ownsAlbum {
+		return nil, errors.New("forbidden")
+	}
+
+	if err := db.Model(&album).Update("cover_id", nil).Error; err != nil {
+		return nil, err
+	}
+
+	return &album, nil
+}
