--- api/graphql/resolvers/resolvers_test.go
@@ -1,12 +1,0 @@
-package resolvers_test
-
-import (
-	"os"
-	"testing"
-
-	"github.com/photoview/photoview/api/test_utils"
-)
-
-func TestMain(m *testing.M) {
-	os.Exit(test_utils.IntegrationTestRun(m))
-}
